requirements.txt
hikari-lightbulb
hikari
asyncio
aiohttp
logging
sqlalchemy
